+++ src/extensions/technotype/technotypeext.cpp
@@ -103,7 +103,8 @@ TechnoTypeClassExtension::TechnoTypeClassExtension(const TechnoTypeClass *this_p
     _JumpjetCloakDetectionRadius(std::numeric_limits<int>::min()),
     JumpjetNoWobbles(false),
     IsNaval(false),
-    BuiltAt()
+    BuiltAt(),
+    BuildTimeMultiplier(1.0f)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("TechnoTypeClassExtension::TechnoTypeClassExtension - Name: %s (0x%08X)\n", Name(), (uintptr_t)(This()));
 }
@@ -379,6 +380,8 @@ bool TechnoTypeClassExtension::Read_INI(CCINIClass &ini)
 
     IdleRate = ini.Get_Int(ini_name, "IdleRate", IdleRate);
     IdleRate = ArtINI.Get_Int(graphic_name, "IdleRate", IdleRate);
+
+    BuildTimeMultiplier = ini.Get_Float(ini_name, "BuildTimeMultiplier", BuildTimeMultiplier);
 
     /**
      *  Fetch the cameo image surface if it exists.