+++ docs/Whats-New.md
@@ -190,6 +190,7 @@ New:
 - Allow up to 65535 OverlayTypes in maps using `NewINIFormat=5` (by ZivDero)
 - Implement naval yards (by ZivDero, CCHyper, Rampastring)
 - Implement exclusive factories (by ZivDero)
+- Change the behavior MultipleFactory and implements build speed overrides (by CCHyper)
 
 
 Vanilla fixes: